pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "skimbleshanks"
authors = [
    {name = "Jeesoo Kim"},
]
description = "Skimbleshanks: The Meticulous Railway Cat for K-Train Reservations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "click",
    "curl_cffi",
    "requests",
    "inquirer>=3.4",
    "keyring",
    "PyCryptodome",
    "prompt_toolkit>=3",
    "python-telegram-bot",
    "fastapi",
    "uvicorn",
    "termcolor"
]
dynamic = ["version"]
[tool.setuptools_scm]

[project.urls]
Homepage = "https://github.com/zits93/skimbleshanks"

[project.scripts]
skimbleshanks = "src.cli.main:skimbleshanks"